+++ resources/views/layouts/app_bottom_nav.blade.php
@@ -16,9 +16,11 @@
                 $currentRouteName = Route::currentRouteName();
             @endphp
             @foreach($bottom_nav_elements as $element)
-                <a href="{{ $element['url'] }}" class="btn text-center p-2 @if($element['active']($currentRouteName)) btn-white text-primary @else btn-white text-muted @endif col" title="{{ $element['description'] }}">
-                    @icon({{ $element['icon'] }})<br>{{ $element['label'] }}
-                </a>
+                @if($element['authorized'])
+                    <a href="{{ $element['url'] }}" class="btn text-center p-2 @if($element['active']($currentRouteName)) btn-white text-primary @else btn-white text-muted @endif col" title="@isset($element['description']){{ $element['description'] }}@endisset">
+                        @icon({{ $element['icon'] }})<br>{{ $element['label'] }}
+                    </a>
+                @endif
             @endforeach
         </div>
     </div>