+++ resources/views/roles/index.blade.php
@@ -1,8 +1,8 @@
-@extends('layouts.app')
+@extends('layouts.user-roles')
 
 @section('title', __('app.roles'))
 
-@section('content')
+@section('wrapped-content')
 
     @if( ! $roles->isEmpty() )
         <table class="table table-sm table-bordered table-striped table-hover">
@@ -10,7 +10,7 @@
                 <tr>
                     <th>@lang('app.name')</th>
                     <th>@lang('app.users')</th>
-                    <th>@lang('app.permissions')</th>
+                    <th class="d-none d-sm-table-cell">@lang('app.permissions')</th>
                 </tr>
             </thead>
             <tbody>
@@ -18,7 +18,7 @@
                     <tr>
                         <td><a href="{{ route('roles.show', $role) }}" title="View Role">{{ $role->name }}</a></td>
                         <td>{{ $role->users->count()  }}</td>
-                        <td>{{ $role->permissions->count()  }}</td>
+                        <td class="d-none d-sm-table-cell">{{ $role->permissions->count()  }}</td>
                     </tr>
                 @endforeach
             </tbody>